Move `Level` into `metadata` module
