+++ tokio-trace-core/src/lib.rs
@@ -78,12 +78,6 @@
 #[macro_use]
 extern crate lazy_static;
 
-use std::borrow::Borrow;
-
-/// Describes the level of verbosity of a `Span` or `Event`.
-#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
-pub struct Level(LevelInner);
-
 pub mod callsite;
 pub mod dispatcher;
 pub mod field;
@@ -97,58 +91,8 @@ pub use self::{
     field::Key,
     span::Id,
     subscriber::{Interest, Subscriber},
-    metadata::Meta,
+    metadata::{Meta, Level},
 };
-
-// ===== impl Level =====
-
-impl Level {
-    /// The "error" level.
-    ///
-    /// Designates very serious errors.
-    pub const ERROR: Level = Level(LevelInner::Error);
-    /// The "warn" level.
-    ///
-    /// Designates hazardous situations.
-    pub const WARN: Level = Level(LevelInner::Warn);
-    /// The "info" level.
-    ///
-    /// Designates useful information.
-    pub const INFO: Level = Level(LevelInner::Info);
-    /// The "debug" level.
-    ///
-    /// Designates lower priority information.
-    pub const DEBUG: Level = Level(LevelInner::Debug);
-    /// The "trace" level.
-    ///
-    /// Designates very low priority, often extremely verbose, information.
-    pub const TRACE: Level = Level(LevelInner::Trace);
-}
-
-#[repr(usize)]
-#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
-enum LevelInner {
-    /// The "error" level.
-    ///
-    /// Designates very serious errors.
-    Error = 1, // This way these line up with the discriminants for LevelFilter below
-    /// The "warn" level.
-    ///
-    /// Designates hazardous situations.
-    Warn,
-    /// The "info" level.
-    ///
-    /// Designates useful information.
-    Info,
-    /// The "debug" level.
-    ///
-    /// Designates lower priority information.
-    Debug,
-    /// The "trace" level.
-    ///
-    /// Designates very low priority, often extremely verbose, information.
-    Trace,
-}
 
 mod sealed {
     pub trait Sealed {}
